only integers

+++ schemas/tourprices.py
@@ -4,11 +4,11 @@ from typing import Optional
 class CreateTourPriceSchema(BaseModel):
     tour_id: int
     currency_id: int
-    price: float
+    price: int
 
 class UpdateTourPriceSchema(BaseModel):
     currency_id: int
-    price: float
+    price: int
 
 
 class TourPriceSchema(CreateTourPriceSchema):
